Improve screenshotting
 - catch errors
 - report snapshots early
